refactor(continuation): merge state files — reply/continuation-state.ts is canonical
continuation/state.ts and reply/continuation-state.ts maintained separate
instances of continuationTimerRefs + continuationTimerHandles Maps (correctness
hazard). Reply-side consumes continuation-side: chain-state persistence
(loadContinuationChainState, persistContinuationChainState) added to the
surviving file. continuationGenerations Map removed per RFC §5.1 (2026-04-15):
generation guard functions kept as no-op stubs for callers.
