+++ src/auto-reply/reply/continuation-state.ts
@@ -1,12 +1,13 @@
+import type { SessionEntry } from "../../config/sessions/types.js";
 import {
   delayedContinuationReservationCount,
   pendingDelegateCount,
   stagedPostCompactionDelegateCount,
 } from "../continuation-delegate-store.js";
+import type { ChainState } from "../continuation/types.js";
 
 type ContinuationTimerHandle = ReturnType<typeof setTimeout>;
 
-const continuationGenerations = new Map<string, number>();
 const continuationTimerRefs = new Map<string, number>();
 const continuationTimerHandles = new Map<string, Set<ContinuationTimerHandle>>();
 
@@ -23,9 +24,8 @@ export function hasDelegatePending(sessionKey: string): boolean {
   );
 }
 
-export function clearDelegatePending(sessionKey: string): void {
-  bumpContinuationGeneration(sessionKey);
-  maybeDropContinuationGeneration(sessionKey);
+export function clearDelegatePending(_sessionKey: string): void {
+  // No-op: pending state is derived from TaskFlow.
 }
 
 export function clearDelegatePendingIfNoDelayedReservations(sessionKey: string): void {
@@ -34,29 +34,21 @@ export function clearDelegatePendingIfNoDelayedReservations(sessionKey: string):
   }
 }
 
-export function currentContinuationGeneration(sessionKey: string): number {
-  return continuationGenerations.get(sessionKey) ?? 0;
+// Generation guard removed per RFC §5.1 (2026-04-15): unrelated channel
+// noise must not cancel dispatched continuation work. Stubs kept for callers.
+export function currentContinuationGeneration(_sessionKey: string): number {
+  return 0;
 }
 
-export function bumpContinuationGeneration(sessionKey: string): number {
-  const next = currentContinuationGeneration(sessionKey) + 1;
-  continuationGenerations.set(sessionKey, next);
-  return next;
+export function bumpContinuationGeneration(_sessionKey: string): number {
+  return 0;
 }
 
 export function hasLiveContinuationTimerRefs(sessionKey: string): boolean {
   return (continuationTimerRefs.get(sessionKey) ?? 0) > 0;
 }
 
-export function maybeDropContinuationGeneration(sessionKey: string): void {
-  if (hasLiveContinuationTimerRefs(sessionKey)) {
-    return;
-  }
-  if (delayedContinuationReservationCount(sessionKey) > 0) {
-    return;
-  }
-  continuationGenerations.delete(sessionKey);
-}
+export function maybeDropContinuationGeneration(_sessionKey: string): void {}
 
 export function retainContinuationTimerRef(sessionKey: string): void {
   continuationTimerRefs.set(sessionKey, (continuationTimerRefs.get(sessionKey) ?? 0) + 1);
@@ -69,7 +61,6 @@ export function releaseContinuationTimerRef(sessionKey: string): void {
   } else {
     continuationTimerRefs.set(sessionKey, current - 1);
   }
-  maybeDropContinuationGeneration(sessionKey);
 }
 
 export function registerContinuationTimerHandle(
@@ -112,4 +103,48 @@ export function clearTrackedContinuationTimers(sessionKey: string): void {
     }, 0);
     releaseHandle.unref();
   }
+}
+
+// ---------------------------------------------------------------------------
+// Chain state persistence (consumed from continuation/state.ts)
+// ---------------------------------------------------------------------------
+
+export type ContinuationChainSource = {
+  continuationChainCount?: number;
+  continuationChainStartedAt?: number;
+  continuationChainTokens?: number;
+};
+
+export function loadContinuationChainState(
+  source: ContinuationChainSource | undefined,
+  turnTokens = 0,
+): ChainState {
+  return {
+    currentChainCount: source?.continuationChainCount ?? 0,
+    chainStartedAt: source?.continuationChainStartedAt ?? Date.now(),
+    accumulatedChainTokens: (source?.continuationChainTokens ?? 0) + turnTokens,
+  };
+}
+
+export function persistContinuationChainState(params: {
+  sessionEntry?: SessionEntry;
+  count: number;
+  startedAt: number;
+  tokens: number;
+}): void {
+  if (!params.sessionEntry) {
+    return;
+  }
+  params.sessionEntry.continuationChainCount = params.count;
+  params.sessionEntry.continuationChainStartedAt = params.startedAt;
+  params.sessionEntry.continuationChainTokens = params.tokens;
+}
+
+// ---------------------------------------------------------------------------
+// Test helpers
+// ---------------------------------------------------------------------------
+
+export function resetContinuationStateForTests(): void {
+  continuationTimerHandles.clear();
+  continuationTimerRefs.clear();
 }
